First Website Push
These are the files I have been working out of up to this point. Some of them are holdovers from the example that we don't really need; I am keeping them for now so that they can be referenced to see how to do stuff, but everything student related will be removed in the final product.